FlaskUI-main/requirements.txt
flask
requests
Pillow
setuptools